+++ odoo/songs/install/data_full.py
@@ -532,6 +532,29 @@ def import_assets(ctx):
 
 
 @anthem.log
+def import_account_assets(ctx):
+    """ Importing account assets from csv"""
+
+    model_asset_category = ctx.env['account.asset.profile'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/account_asset_profile.csv',
+             model_asset_category)
+
+    model = ctx.env['account.asset'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/account_asset.csv',
+             model)
+    model_line = ctx.env['account.asset.line'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx,
+             's3://prod-sf-odoo-data/install/account_asset_line.csv',
+             model_line)
+
+
+@anthem.log
 def import_serial_number_stock_moves(ctx):
     """ Importing stock_moves with serial number from csv """
     pick_id = ctx.env['stock.picking'].create(